Librarify tidy
Convert tidy into a library so that the data it creates
can be used by external tools.
